==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Steps for Downloading and Saving the data from U.S. GSS
# Author: Gauravpreet Thind
# Date: 8th Nov 2024
# Contact: gauravpreet.thind@mail.utoronto.ca
# License: MIT
# Pre-requisites: N/A

#### Workspace setup ####
# Not required

#### Download data ####

# The download of the raw data requires going to the GSS website and creating an account. 

# Step 1: Visit the Website

# The GSS data can be accessed through the NORC at the University of Chicago.
# Start by visiting the GSS website: https://gssdataexplorer.norc.org/MyGSS.


# Step 2: Register

# To download the GSS data, you are required to register an account on the website.
# Follow the registration instructions provided on the GSS website.


# Step 3: Locate the Relevant Variables to Extract from the GSS dataset

# Once registered, you can access the datasets.
# The GSS website provides a search and filter function to help you find specific datasets.
# Select the variable names "YEAR", "SEX", "AGE", "HEALTH", "PHYSHLTH", "MNTLHLTH", "DEPRESS" 
# for extracting.


# Step 4: Download the Data

# After locating the relevant variables, the dataset can be downloaded. 
# The GSS website offers data in various formats, Choose Excel Workbook (data + metadata) for replicating.

#### Save data ####

# No further action is required to saving data
